scripts/setup.py: fix for Fedora 43
Add support for Fedora 43 in scripts/setup.py. Fedora 43 no longer has
a "java-1.8.0-openjdk" package, so we need to remove it from the list
of dependencies (and I had to add a feature "removed" to be able to
say that Fedora 43 removed some package from the list of dependencies).

Also fixed a few Python warnings.

I'm happy to report that no other changes are needed to get OSv to build
or run on Fedora 43. I checked the default build (Lua shell), "rogue"
and "tests" (and scripts/test.py), and they all build and run fine.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>